+++ src/main/java/com/soin/sgrm/service/SigesService.java
@@ -5,5 +5,5 @@ import java.util.List;
 import com.soin.sgrm.model.Siges;
 
 public interface SigesService extends BaseService<Long, Siges>{
-	public List<Siges> listCodeSiges(Long id);
+	public List<Siges> listCodeSiges(Integer id);
 }
